third_party/blink/web_tests/external/wpt/css/css-shapes/shape-functions/shape-function-valid.tentative-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['clip-path'] = "shape(from 20px 40px, curve to right center with 10px 30px from  end / 12px 32px from  origin)" should set the property value
  assert_equals: serialization should be canonical expected "shape(from 20px 40px, curve to right center with 10px 30px from end / 12px 32px from origin)" but got "shape(from 20px 40px, curve to right center with 10px 30px from end / 12px 32px)"
[FAIL] e.style['clip-path'] = "shape(from 20px 40px, curve to right center with 10px 30px from origin / 12px 32px from start)" should set the property value
  assert_equals: serialization should be canonical expected "shape(from 20px 40px, curve to right center with 10px 30px from origin / 12px 32px from start)" but got "shape(from 20px 40px, curve to right center with 10px 30px / 12px 32px from start)"
[FAIL] e.style['clip-path'] = "shape(from 20px 40px, curve by 20% 20em with 10px 30px from start / 12px 32px from end)" should set the property value
  assert_equals: serialization should be canonical expected "shape(from 20px 40px, curve by 20% 20em with 10px 30px from start / 12px 32px from end)" but got "shape(from 20px 40px, curve by 20% 20em with 10px 30px / 12px 32px from end)"
[FAIL] e.style['clip-path'] = "shape(from 20px 40px, curve by 20% 20em with 10.3% 30px from origin / 12pt 5.4% from start)" should set the property value
  assert_equals: serialization should be canonical expected "shape(from 20px 40px, curve by 20% 20em with 10.3% 30px from origin / 12pt 5.4% from start)" but got "shape(from 20px 40px, curve by 20% 20em with 10.3% 30px from origin / 12pt 5.4%)"
[FAIL] e.style['clip-path'] = "shape(from 20px 40px, smooth by 20pt 20px with 12px 32px from start)" should set the property value
  assert_equals: serialization should be canonical expected "shape(from 20px 40px, smooth by 20pt 20px with 12px 32px from start)" but got "shape(from 20px 40px, smooth by 20pt 20px with 12px 32px)"
Harness: the test ran to completion.
